[various] fix PID grab and allow for custom payload dirs
On NixOS, process names cant be reliably grabbed using pgrep or by
grepping ps. Instead we use the built-in BASH functionality for that,
despite what my previous comment may suggest about it's reliability.
Hopefully it doesn't break it.

The custom payload dirs thing was done to allow for potential packaging
of this application, in case it needs to be done.